Speed up build workflow with ccache
Use ccache to speed up C and C++ compilation triggered by build.yml.
This only affect macOs and Linux runners.

For macOs, installation and configuration and github caching is done
through the hendrikmuhs/ccache action.
Cmake is set to use ccache through the env variable CMAKE_*_COMPILER_LAUNCHER.

For Linux, installation is done in build-skdecide_dev, config and link with
Cmake in build-many-linux-wheels.sh.
The cache directory is fetched from the docker image to the .ccache directory
to be cached by the default github cache action and reused in later runs

For Linux, the setup is gated by the existence of .ccache directory that is either
restored from github cache or created in Build_x86_64. This ensure that release.yml
doesn't use ccache while still sharing the dockerfiles and scripts used by build.yml

For both runners, ccache is configured to use compression (by default for the version used by macOs)
and limits cache size to 80M